fix: performed multiple fixes reverted speaker_verif to 41fdf34 changer order in subset method creation moved load_trial & load_trial to loader.py
